api/serializers/questionnaire.py
from rest_framework import serializers
from api.models import Questionnaire


class QuestionnaireSerializer(serializers.ModelSerializer):
    class Meta:
        model = Questionnaire
        fields = '__all__'
